+++ vllm_ascend/patch/worker/patch_v2/patch_eagle.py
