+++ vllm_ascend/worker/v2/utils.py
@@ -5,7 +5,6 @@ import vllm
 from vllm.logger import logger
 
 from vllm_ascend.worker.v2.block_table import AscendBlockTables
-from vllm_ascend.worker.v2.model_states import init_asecnd_model_state
 
 
 @contextmanager
@@ -24,30 +23,6 @@ def torch_cuda_wrapper():
         torch.cuda.current_device = torch.npu.current_device
         torch.cuda.mem_get_info = torch.npu.mem_get_info
         logger.info_once("Wrapping torch.cuda with torch.npu.")
-        yield
-    finally:
-        pass
-
-
-@contextmanager
-def block_table_wrapper():
-    try:
-        # vllm-ascend need to initialize slot mapping as torch.int32 dtype,
-        # but vllm default is torch.int64 dtype.
-        vllm.v1.worker.gpu.model_runner.BlockTables = AscendBlockTables
-        logger.info_once("Wrapping BlockTables with AscendBlockTables.")
-        yield
-    finally:
-        pass
-
-
-@contextmanager
-def model_states_wrapper():
-    try:
-        # prepare_attn in AscendModelState is different from vllm,
-        # we need to override init_model_state.
-        vllm.v1.worker.gpu.model_runner.init_model_state = init_asecnd_model_state
-        logger.info_once("Wrapping init_model_state with init_asecnd_model_state.")
         yield
     finally:
         pass
